Track.py: - Added a class called Track. This will be responsible for storing the cone points and estimated car route after mapping has occurred. - Also in here is where racing-line optimisations can be performed.
Detections.py:
- Added loc_ws property to DetectedCone
- Added unmapped property (default value is True) to DetectedCone. This
will be set to false when the cone is mapped for the first time. Then
when a cone from a subsequent cone is matched to a cone that has already
been mapped, unmapped will be set to false to ensure that the same cone
is not mapped multiple times.

Mapping.py:
- Added a few useful methods to point class

StereoMatching.py:
- Added functionality to set DetectedCone.unmapped = False if a cone is
matched to a cone that has already been mapped.